kvserver: move syncing_write.go to kvserverbase
This file is used both by log storage (storing AddSST commands in sideloaded
storage) and state storage (applying AddSST). Moving it to kvserverbase to
avoid circular dependencies when factoring out the log storage.

Release note: None
