+++ pkg/kv/kvserver/BUILD.bazel
@@ -94,7 +94,6 @@ go_library(
         "stores.go",
         "stores_base.go",
         "stores_server.go",
-        "syncing_write.go",
         "testing_knobs.go",
         "track_raft_protos.go",
         "ts_maintenance_queue.go",
