+++ pkg/kv/kvserver/kvserverbase/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "knobs.go",
         "raftversion.go",
         "stores.go",
+        "syncing_write.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverbase",
     visibility = ["//visibility:public"],
@@ -19,14 +20,18 @@ go_library(
         "//pkg/kv/kvserver/kvserverpb",
         "//pkg/roachpb",
         "//pkg/settings",
+        "//pkg/settings/cluster",
+        "//pkg/storage/fs",
         "//pkg/util/errorutil",
         "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
         "//pkg/util/log",
         "//pkg/util/quotapool",
+        "//pkg/util/timeutil",
         "//pkg/util/tracing/tracingpb",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_redact//:redact",
+        "@org_golang_x_time//rate",
     ],
 )
